WIP before cleaning and adding domainFileStore check; fixes #110
