+++ IntegrationTestSuite/WebservicesIntegrationTests/Tests/EngineeringModel/DomainFileStore/DomainFileStoreTestFixture.cs
@@ -118,88 +118,5 @@ namespace WebservicesIntegrationTests
             IList<string> folders = foldersArray.Select(x => (string) x).ToList();
             CollectionAssert.AreEquivalent(expectedFolders, folders);
         }
-
-        [Test]
-        public void VerifyThatDomainFileStoreWillNotBeDeletedWhenDomainOfExpertiseWillBeRemovedFromExistingEngineeringModelSetup()
-        {
-            var engineeringModelSetupsUri =
-                new Uri(string.Format(UriFormat, this.Settings.Hostname,
-                    "/SiteDirectory/f13de6f8-b03a-46e7-a492-53b2f260f294/model"));
-
-            var jArray = this.WebClient.GetDto(engineeringModelSetupsUri);
-            Assert.AreEqual(1, jArray.Count);
-            var engineeringModelSetup = jArray.Single(x => (string) x[PropertyNames.Iid] == "116f6253-89bb-47d4-aa24-d11d197e43c9");
-            Assert.AreEqual("116f6253-89bb-47d4-aa24-d11d197e43c9", (string) engineeringModelSetup[PropertyNames.Iid]);
-            var model = (string) engineeringModelSetup[PropertyNames.EngineeringModelIid];
-            Assert.AreEqual("9ec982e4-ef72-4953-aa85-b158a95d8d56", model);
-
-            // Check DomainOfExpertise in EngineeringModelSetup
-            var expectedActiveDomains = new string[]
-            {
-                "0e92edde-fdff-41db-9b1d-f2e484f12535",
-                "eb759723-14b9-49f4-8611-544d037bb764"
-            };
-
-            var activeDomainsArray = (JArray) engineeringModelSetup[PropertyNames.ActiveDomain];
-            IList<string> activeDomainsList = activeDomainsArray.Select(x => (string) x).ToList();
-            CollectionAssert.AreEquivalent(expectedActiveDomains, activeDomainsList);
-
-            // Check DomainFileStore in EngineeringModel correlated to EngineeringModelSetup
-            var engineeringModelUri =
-                new Uri(string.Format(UriFormat, this.Settings.Hostname,
-                    "/EngineeringModel/" + model));
-
-            jArray = this.WebClient.GetDto(engineeringModelUri);
-            var engineeringModel = jArray.Single(x => (string) x[PropertyNames.Iid] == model.ToString());
-            var iterationsArray = (JArray) engineeringModel[PropertyNames.Iteration];
-
-            var iterationUri =
-                new Uri(string.Format(UriFormat, this.Settings.Hostname,
-                    "/EngineeringModel/9ec982e4-ef72-4953-aa85-b158a95d8d56/iteration/" + iterationsArray[0].ToString()));
-
-            jArray = this.WebClient.GetDto(iterationUri);
-            var iteration = jArray.Single(x => (string) x[PropertyNames.Iid] == iterationsArray[0].ToString());
-            var domainFileStoresArray = (JArray) iteration[PropertyNames.DomainFileStore];
-            Assert.AreEqual(1, domainFileStoresArray.Count);
-
-            // Check DomainFileStore after delete DomainOfExpertise in EngineeringModelSetup
-            var siteDirectoryUri =
-                new Uri(string.Format(UriFormat, this.Settings.Hostname,
-                    "/SiteDirectory/f13de6f8-b03a-46e7-a492-53b2f260f294"));
-
-            var postBodyPath = this.GetPath("Tests/EngineeringModel/DomainFileStore/PostDeleteDomainOfExpertise.json");
-            var postBody = base.GetJsonFromFile(postBodyPath);
-            jArray = this.WebClient.PostDto(siteDirectoryUri, postBody);
-
-            engineeringModelSetupsUri =
-                new Uri(string.Format(UriFormat, this.Settings.Hostname,
-                    "/SiteDirectory/f13de6f8-b03a-46e7-a492-53b2f260f294/model/116f6253-89bb-47d4-aa24-d11d197e43c9"));
-
-            jArray = this.WebClient.GetDto(engineeringModelSetupsUri);
-            engineeringModelSetup = jArray.Single(x => (string) x[PropertyNames.Iid] == "116f6253-89bb-47d4-aa24-d11d197e43c9");
-            expectedActiveDomains = new string[] { };
-            activeDomainsArray = (JArray) engineeringModelSetup[PropertyNames.ActiveDomain];
-            activeDomainsList = activeDomainsArray.Select(x => (string) x).ToList();
-            CollectionAssert.AreEquivalent(expectedActiveDomains, activeDomainsList);
-
-            jArray = this.WebClient.GetDto(iterationUri);
-            iteration = jArray.Single(x => (string) x[PropertyNames.Iid] == iterationsArray[0].ToString());
-            domainFileStoresArray = (JArray) iteration[PropertyNames.DomainFileStore];
-            Assert.AreEqual(1, domainFileStoresArray.Count);
-
-            // Check DomainFileStore after adding DomainOfExpertise in EngineeringModelSetup
-            siteDirectoryUri =
-                new Uri(string.Format(UriFormat, this.Settings.Hostname,
-                    "/SiteDirectory/f13de6f8-b03a-46e7-a492-53b2f260f294"));
-
-            postBodyPath = this.GetPath("Tests/EngineeringModel/DomainFileStore/PostAddDomainOfExpertise.json");
-            postBody = base.GetJsonFromFile(postBodyPath);
-            jArray = this.WebClient.PostDto(siteDirectoryUri, postBody);
-
-            jArray = this.WebClient.GetDto(iterationUri);
-            iteration = jArray.Single(x => (string) x[PropertyNames.Iid] == iterationsArray[0].ToString());
-            domainFileStoresArray = (JArray) iteration[PropertyNames.DomainFileStore];
-            Assert.AreEqual(1, domainFileStoresArray.Count);
-        }
     }
 }
