+++ IntegrationTestSuite/WebservicesIntegrationTests/Tests/SiteDirectory/EngineeringModelSetup/PostDeleteActiveDomain.json
@@ -1,0 +1,11 @@
+{
+  "_delete": [
+    {
+      "activeDomain": [ "0e92edde-fdff-41db-9b1d-f2e484f12535", "eb759723-14b9-49f4-8611-544d037bb764" ],
+      "classKind": "EngineeringModelSetup",
+      "iid": "ba097bf8-c916-4134-8471-4a1eb4efb2f7"
+    }
+  ],
+  "_create": [],
+  "_update": []
+}
