+++ IntegrationTestSuite/WebservicesIntegrationTests/Tests/SiteDirectory/EngineeringModelSetup/PostUpdateIteration.json
@@ -1,0 +1,12 @@
+{
+  "_delete": [],
+  "_create": [],
+  "_update": [
+    {
+      "domainFileStore": [ "da7dddaa-02aa-4897-9935-e8d66c811a96" ],
+      "classKind": "Iteration",
+      "iid": "!"
+    }
+  ],
+  "_copy": []
+}
